# docs/swagger.yaml
definitions:
  model.Dummy:
    properties:
      id:
        type: integer
      info:
        type: string
    type: object
  request.Dummy:
    properties:
      info:
        type: string
    required:
    - info
    type: object
  response.PaginationResponse:
    properties:
      pageNumber:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  response.ResponseError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.RestResponsePagination-model_Dummy:
    properties:
      elements:
        items:
          $ref: '#/definitions/model.Dummy'
        type: array
      pagination:
        $ref: '#/definitions/response.PaginationResponse'
    type: object
  response.restResponseError:
    properties:
      errors:
        items:
          $ref: '#/definitions/response.ResponseError'
        type: array
    type: object
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /dummy:
    get:
      consumes:
      - application/json
      description: Get a list of dummies with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Size per page
        in: query
        name: size
        type: integer
      - description: Sort by field
        in: query
        name: sortBy
        type: string
      - description: Sort order (asc or desc)
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of dummies
          schema:
            $ref: '#/definitions/response.RestResponsePagination-model_Dummy'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/response.restResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.restResponseError'
      security:
      - OAuth2Password: []
      summary: List all dummies
      tags:
      - dummy
    post:
      consumes:
      - application/json
      description: Create a new dummy item with the provided information
      parameters:
      - description: Dummy information
        in: body
        name: dummy
        required: true
        schema:
          $ref: '#/definitions/request.Dummy'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Dummy'
        "400":
          description: Invalid request body or validation errors
          schema:
            $ref: '#/definitions/response.restResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.restResponseError'
      security:
      - OAuth2Password: []
      summary: Create a new dummy item
      tags:
      - dummy
  /dummy/{id}:
    get:
      consumes:
      - application/json
      description: Get dummy details by ID
      parameters:
      - description: Dummy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Dummy'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/response.restResponseError'
        "404":
          description: Internal Error
          schema:
            $ref: '#/definitions/response.restResponseError'
      security:
      - OAuth2Password: []
      summary: Find a dummy by ID
      tags:
      - dummy
  /dummy/external/api:
    get:
      consumes:
      - application/json
      description: Call external API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Dummy'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/response.restResponseError'
        "404":
          description: Internal Error
          schema:
            $ref: '#/definitions/response.restResponseError'
      security:
      - OAuth2Password: []
      summary: Call external API
      tags:
      - dummy
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: KEYCLOAK_HOST/realms/javi/protocol/openid-connect/token
    type: oauth2
swagger: "2.0"